requirements.txt
blinker==1.9.0
cachelib==0.9.0
click==8.1.8
Flask==3.1.0
Flask-Caching==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
pip==24.3.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
Werkzeug==3.1.3